extract encoding module
