+++ modules/encoding/path/path.replacements.ts
