+++ modules/encoding/session/session.stringify.ts
@@ -1,4 +1,4 @@
-import { AlgolLocalBattle } from "../../../../types";
+import { AlgolLocalBattle } from "../../types";
 
 export const stringifySession = (local: AlgolLocalBattle, method: number) => {
   if (method === 0) {
