--- modules/local/src/session/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from "./pathReplacements";
