--- modules/local/src/session/parseSession/index.ts
@@ -1,2 +1,0 @@
-export * from "./parseSession";
-export * from "./parseBattleSave";
