--- modules/local/src/session/stringifySession/index.ts
@@ -1,2 +1,0 @@
-export * from "./stringifySession";
-export * from "./stringifyBattleSave";
